Revert atomic linked list and use a mutex instead for deferred tri freeing.
There is a subtle issue in the deferred tri which depends on the order
of deletion (because some tris reference data from other tris), and since the lock-free list basically reverts the order,
we'd have to inverse the list again here. Not worth it, since the mutex
doesn't appear to be any slower, anyway.