#5992. Allow AIs to think only on the first gametic per frame (controlled by com_useMinorTics).
Most of the time in game modelling is taken by alive AIs, so disabling them is a good way to accelerate it and avoid spiral of death.
Luckily, they already skip thinking when player does not see them, thanks to Interleaved Thinking Optimization.

So, if com_useMinorTics is 1, then all game tics except for the first one are marked as "minor".
idGameLocal::RunFrame saves this flag into self, and then we check it in idAI::ThinkingIsAllowed.

In idAI::ThinkingIsAllowed function, added a major check that we never skip thinking for more than what GetMaxInterleaveThinkFrames allows.
So if we have extra-low FPS and very many minor tics, AIs will think on some of the minor tics to avoid issues with too rare thinking.
How to refactor some code here BTW.



git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@9981 49c82d7f-2e2a-0410-a16f-ae8f201b507f