#5869. idWinVar::Set now returns false in case of wrong format, instead of posting a warning immediately.
Additionally:
1) Now idWinBool only considers 0 and 1 as valid values, other integers mean bad format.
2) Now we check that input string is read up to the end, so e.g. "1st" and "0.7" are now wrong integers.

Note that the main change (don't warn) means reducing our diagnostics in the short term.
The plan is to wrap all calls to idWinVar::Set into a check, and report warning message there.
These changes will be in the future commits.

This would be much better for two reasons:
1) We have more context at call site, e.g. we can print source code location and window instead of simply the value which we don't like.
2) We can be sure we won't spam warning continuously at runtime, since we can only warn during static analysis.


git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10015 49c82d7f-2e2a-0410-a16f-ae8f201b507f