#5869. Don't enable string concatenation by backslash (\) in GUI code.
GUI code uses doublequotes everywhere to make names like gui::my_var atomic, since otherwise parser would break them into many tokens.
For this reason, string concatenation is disabled in lexer (that's when you can write "hello" " " "world", and C will treat it as single string "hello world").

However, ID enabled special string concatenation via backslash to substitute for it, like this:
  "hello" \
  " " \
  "world"
The problem with such usage is that it breaks multiline macros.

Due to limitations of GUI code, we have to write whole window (even with subwindows) into a macro in order to make it reusable.
If any line in such macro ends with a string literal (that's very likely in GUI code), then parsing breaks and whole macro does not work.
This can be worked around by moving the first token on the next line to the end of the current line, but it is very messy and totally not obvious.
Better just forbid string concatenation altogether.

Buy the way, the following way of concatentating string works (at least inside macro):
  "hello" ## " " ## "world"
Note that ## is a standard C way of concatenating tokens, although in C it works on TOKENS, not on string literals --- but here they are mixed anyway.
Also, it is even possible to do this:
  #define HELLO_MESSAGE(index) "Hello, " ## #idx ## "-th user!"
Which is very useful to construct variable names inside macros =)


git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10031 49c82d7f-2e2a-0410-a16f-ae8f201b507f