#6028. Fixed awful bug in getting component of vector in GUI script.
A pointer "idWinVec4 *var" was casted to reference "idVec4&" with C-style cast.
Of course, the author forgot to dereference the pointer in order to later rely on custom conversion operator (damn).
As the result, v[0] and v[1] are float-reintepreted halves of "var"'s address, and v[2] and v[3] are float-read contents of subsequence stack space =)


git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10049 49c82d7f-2e2a-0410-a16f-ae8f201b507f