Deleted DECL_TDM_MATINFO type of decl file and all accompanying code.
It was implemented back in 2006, and it not used anywhere.
There is no C++ code that uses the information that might get parsed from such decls.
There is no decl file in core or in FMs that define tdm_matinfo.

Also it is nasty because it parses .mtr files in materials directory (exactly like actual materials do).
So if any .mtr file does not set keyword "material", then it will get parsed by this damn code...
Such a situation when same files get parsed several times blocks me from supporting FM > core decl overrides (#5766).

Obviously, setting type of surface by material is already implemented in a different way, so there is no reason to keep this.


git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10059 49c82d7f-2e2a-0410-a16f-ae8f201b507f