#5828 !!! Now specular term is modulated by specular color instead of diffuse color in ambient shader.
Note that in #5044 we did the same change for interaction shader: it used diffuse color, and now uses specular color.
Despite big expectations for all kind of issues, the change worked smoothly.
Now we should make ambient shader behavior the same.

This changes the way things look, so might be a candidate for revert in future...

git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10088 49c82d7f-2e2a-0410-a16f-ae8f201b507f