#5887. Raised tracemodel limits: polys[16 -> 32], edges[32 -> 63], verts[32].
For a closed mesh consisting of d-gons:
  d = 3 : up to 32 triangles
  d = 4 : up to 31 quads
  d = 5 : up to 20 pentagons

Note that cm_vertex_t and cm_edge_t both contain members "side" and "sideSet", which are bitsets.
Their size is proportional to number of vertices/edges/polys in tracemodel (depending on whether it's point check or translation check).
So I had to expand them from 32 bits to 64 bits, as well as all bitmask manipulation.
I'm afraid 63 edges is maximum that we can process this way.


git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10143 49c82d7f-2e2a-0410-a16f-ae8f201b507f