#5886. Reimplemented shadow volume computation with BVH by storing all triangles in a bitset.
During shadow volume computation, we need to store status flag for every triangle: is it shadow-casting or not?
The old BVH code stores the flags only for the intervals which were found by BVH.
While this approach allows to NOT use O(N) memory in strictest sense, it gives quite a lot of overhead in practice.

The new BVH code stores statuses for all triangles again to reduce overhead.
But there are a few tricks to reduce memory bandwidth:
1) Flags are stored in idBitArray, i.e. they take 1 bit per triangle instead of 1 byte per triangle (more cache-friendly).
2) We touch the flags only on the intervals returned to us by BVH traversal, don't touch all flags.

The p.2 is rather fragile.
Normally, when we initialize a bitset (of size = #(tris)), we must set all bits to zero.
But this takes O(N) execution time, which is quite slow on huge models and kills the idea of BVH optimization.
Instead, we reuse the same bitset container between subsequent calls, and cleanup it after every use (only on the intervals which we could modify).
This allows us to only use O(K) memory where K is sum of lengths of all intervals returned by BVH, keeping bitset clean between uses.
Of course, the bitset has to be threadlocal, and grow exponentially when its size is not enough.

The much simpler code for accessing shadow-casting flags allowed to do a few low-level optimizations:
1) Optimized filling the statuses with SSE + bitwise operations.
2) Unroll the code for checking three edges of a triangle.
3) Optimized 6 AddGrow calls into single SetNum.


git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10174 49c82d7f-2e2a-0410-a16f-ae8f201b507f