TiledCustomMipmapStage: set full scissor at the end, instead of calling FB_ApplyScissor to restore previous one.
The problem here is that FB_ApplyScissor does nothing when r_useScissor = 0.
Which normally should never happen... except it is disabled for screenshots for whatever crazy reason =)

So in case of r_useScissor = 0, the class did change glScissor, but did not restore its original value.
Now we restore full scissor: the subsequent code should reapply some new scissor anyway.
In the worse case we'll lose some bandwidth, which is not as big as buggy rendering.

It is also questionable: should TiledCustomMipmapStage respect r_useScissor and not change scissor if cvar is 0?
I'd say better not, because otherwise full-screen passes will become much more costly, making this optimization more like "pessimization".

This change fixes the issue of black screenshots:
  https://forums.thedarkmod.com/index.php?/topic/21662-timing-of-211-beta/&do=findComment&comment=480815

git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10187 49c82d7f-2e2a-0410-a16f-ae8f201b507f