Don't disable r_useScissor for screenshots.
This commit fixes the issue:
  https://forums.thedarkmod.com/index.php?/topic/21679-beta-testing-211/#comment-480979

Unfortunately, the game does not work properly now with scissors disabled.
That's because r_useScissor causes some code to not change scissor state, while some other code still sets scissor unconditionally.
The result is obviously scissor state leak, which is perhaps easy to fix once but hard to keep fixed always.
It is much better to push such an important feature as screenshots onto a well-tested and common path.

Tiled screenshots don't work anyway.
It is such a niche feature that it is not worth to sacrifice stability of the main feature for it.


git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10198 49c82d7f-2e2a-0410-a16f-ae8f201b507f