#6223. Wait at least one full period before firing a periodic stim.
Before this commit, a stim on a spawned entity first triggers after random delay t in [0..T], where T is sr_interval.
In particular, it can spawn immediately (well, maybe one frame later).

This commit returns the delay almost like it was before revision 9910 "Improve distribution of stims with same period over frames".
In particular, it usually fixes the issue of periodic stim being generated before script object is initialized.
This issue causes problems for lights, and this fix is quite necessary for custom light entityDefs, although it is not that required for core ones (see svn rev 16748).


git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10250 49c82d7f-2e2a-0410-a16f-ae8f201b507f