#6223. Delay stims one full period from game start, not from entity spawning time.
It turned out that water arrow effect is also a periodic stim with 100ms period (WHY?)
So if we delay it by one period after spawning, then direct hit to a torch does not dose it.

Also, this change is closer to the state of 2.10 (i.e. before svn rev 9910).
There m_TimeInterleave was always left to 0, so it only caused full-period delay on map start, but now any later.

git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10251 49c82d7f-2e2a-0410-a16f-ae8f201b507f