#6264. Improved subtitles system in several ways.
1) Fixed some weird handling of looping sounds in idSoundChannel::GatherSubtitles.
Verbosity can be different for leadin and loop, also we should fetch subtitles from loop sample when it is playing.

2) Now more information is returned for subtitle: + sound emitter, sound sample, verbosity, 

3) Better algorithm for assignment of subtitles to slots.
First of all, high-level subtitles are always displayed in case of overflow, they are never dropped due to many low-level subtitles.
Second, I fixed the bug that subtitle could be moved (even without overflow) if some new subtitles was added to the list of matches.
Also, now the algorithm uses the new grouping info (sample and emitter) for better reusage of slots.

4) Set number of slots to 3 as it is in GUI.
If we pretend we have more slots that there really are, then slots assignment rules can't do the best assignment in case of overflow.


git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10285 49c82d7f-2e2a-0410-a16f-ae8f201b507f