#6271. Support alpha-tested shadow mapping under r_shadowMapAlphaTested cvar.
Currently it only affects "r_useNewBackend 1" + "r_shadowMapSinglePass 1", but in future this cvar will affect all surviving modes.
If r_shadowMapAlphaTested is 0 (default), then alpha-test is ignored and whole surfaces casts shadows --- just like with stencil shadows.

Note: the reason for adding this support now is that there was already some code related to it but commented out.
Better uncomment and fix it now.
Also, the old backend always works like that always.

git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10306 49c82d7f-2e2a-0410-a16f-ae8f201b507f