#6271. Added blue noise dithering to tonemapping shader on 64-bit color precision.
This is used to fix color banding of fog, for instance on March of Rahena mission (easy to notice when you fly to the castle immediately after start).
Requires r16797 "Added 1024 x 1024 image with tileable RGBA blue noise" from assets repo to work.

The blue noise image is loaded from assets.
During tonemapping, we add some dithering before the tonemapping formula and after it.
Note that dithering before tonemapping is most likely useless, since the higher-precision information is already lost at that point.

For the same reason, dithering is force-disabled on 32-bit or 16-bit FBO colors: it does not help.
Because the high-precision information was lost before tonemapping, e.g. when we used fog shader of interaction shader.
I think that's not a big loss, since 64-bit color is default, and using lower only makes sense for some kind of "nostalgia for shitty graphics" reason =)

This does not solve the problem completely, but reduces it greatly.
I'm afraid the problem can still happen depending on scene and gamma + brightness settings.

git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10370 49c82d7f-2e2a-0410-a16f-ae8f201b507f