idBitArray: fixed memory overallocation, added growth capabilities.
It turns out that the implementation of bitset allocates numBits WORDS of memory, instead of numBits BITS of memory.
So it overallocates in 64 times =(
While it is probably not a big issue, method SetBitsSameAll also fills num WORDS instead of num BITS, i.e. touches 64x amount of memory.

Also, I added AddGrow method, which allows to increase size gradually (with typical exponential growth).
This is required for the cases when size is not known beforehand.

git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10383 49c82d7f-2e2a-0410-a16f-ae8f201b507f