#6296. Store bitset in render world, which duplicate information of whether entity is in current view or not.
CreateLightDefInteractions contains the following important optimization:
  // if the entity isn't viewed and light has now shadows, skip
  if ( tr.viewDef && !lightCastsShadows && edef->viewCount != tr.viewCount )
    continue;
It allows to greatly reduce amount of work for noshadows lights.

The problem here is that in order to do this check, we need to touch idRenderEntityLocal::viewCount, which is very expensive.
Now we duplicate this information in a bitset idRenderWorldLocal::entityDefsInView, which is indexed by entityIdx.
This allows to check whether entity is in view without touching entity itself.
In fact, we only need random access to the bitset, which is e.g. 1.5 KB for 12K entities (that's about the maximum we evet have now) and should fit L1 cache nicely.
This reduces the CPU fraction to about 5.5% on Iris start from about 7-8% initial (the previous commit gives wrong perfor numbers, sorry).

Also I had to refactor addition of render entity into AllocateEntityDefHandle method.
Previously it was done in several places, but now I need to synchronously grow the bitset when entity is added.


git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10385 49c82d7f-2e2a-0410-a16f-ae8f201b507f