#6296. Pass indices of light + entity instead of pointers to idInteractionTable::Find.
Note that hash table key is composed of light and entity indices.
Previously we passed pointers to objects, then read indices from the objects.
It means that we had to touch idRenderEntityLocal, which is costly.

Now we simply pass index which is directly available.
Touching entity is no longer needed e.g. when this interaction was previously considered and is empty.

git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10388 49c82d7f-2e2a-0410-a16f-ae8f201b507f