#6296. Don't call R_SetEntityDefViewEntity if entity is in view.
In interaction table request hits, then we normally don't have to do anything.
However, R_SetEntityDefViewEntity is called to force creation of viewEntity for the entity.
Internally, it exits if entity is already marked as "in view", but this check is done via "entityDef->viewCount == tr.viewCount", which touched entityDef and is thus expensive.
Now we have this value from bitset, so we can do a cheap check outside.

The time for CreateLightDefInteractions is reduced to almost 2% =)

git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10390 49c82d7f-2e2a-0410-a16f-ae8f201b507f