dmap ear-cutting: add "desperate mode" to always produce topologically correct triangulation.
In the rare case when there are no ears to cut, previously algorithm stopped.
The remaining portion of the polygon was left unfilled, and the triangulation was not topologically equivalent to the input polygon.

Now we use so-called "desperate mode" from FIST triangulation article.
I.e. we simply pick any of the remaining vertices and cut it even if it is not considered ear.
This allows algorithm to continue and always produce triangulation which is topologically correct, even though geometrically it might have self-overlappings and self-intersections.

Note: as of now, I get the warning only in Painter's Wife, and there this hack does not help to fix the decal =(
But I think it should be helpful anyway.

git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10398 49c82d7f-2e2a-0410-a16f-ae8f201b507f