#6224. Deleted references of light_loss spawnarg on portal entities.
Its description reads "Lightprop: Loss incurred when light travels through this portal. (Not implemented)".
This code was added in 2010, and there is no implementation still.

In fact, implementation is hardly possible, because an entity can easily cross visportal, in which case it is not clear how to apply this light loss.
Lighting entity twice with different clipping plane is a huge mess and would look awful, applying light loss based on origin is hardly any better.

git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10400 49c82d7f-2e2a-0410-a16f-ae8f201b507f