#6300. Deleted Ftou function completely.
It is very niche function, used only in two places.
Most likely it can be replaced with Ftoi in both cases...
It is also a bit dangerous, since passing negative value would cause unexpected result.
It is deleted in Doom 3 BFG too, which confirms it's uselessness.

In the two usages, I replaced the call with cast to unsigned int, so there is no behavior change here.

git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10412 49c82d7f-2e2a-0410-a16f-ae8f201b507f