#6306. Disabled connectivity check for parallelSky lights in idRenderWorldLocal::AddAreaLightRefs (controlled by r_useLightAreaCulling).
This one is totally wrong, since it checks if the area containing the light "origin" is connected to the area being lit.
It was disabled when origin is outside playable areas, maybe that's what saved it.

Note: I don't disable this for ordinary parallel lights, since they are in "locked" state now --- no changes in behavior without breaking hacky missions.


git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10430 49c82d7f-2e2a-0410-a16f-ae8f201b507f