Fixed invalid-offsetof in the code + set C++ standard only for C++ sources (don't set for C sources).
Setting C++14 and invalid-offsetof for C sources results in unknown parameter warning.
So now we set C++14 only for C++ sources using cmake generator expression.

As for invalid-offsetof, we no longer suppress this warning globally.
I tried to suppress it locally with #pragma GCC diagnostic, but for some reason it does not work even if applied on the whole function.
So instead I use #pragma pack to ensure proper 8-byte prefix before payload.
There is a bunch of asserts and static_asserts to check that this is correct.
Note that now overaligned (alignof > 8) types cannot be put into idBlockAlloc, that would produce compile error.


git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10575 49c82d7f-2e2a-0410-a16f-ae8f201b507f