#5172. Set forceShadowBehindOpaque on all doors with visportal connected.
This fixes the issue in "Poets and Peasants":
  https://forums.thedarkmod.com/index.php?/topic/22241-new-shadow-leaks-in-212-dev-builds/&do=findComment&comment=492137

Note that if door is connected to visportal, then it becomes part of the wall and blocks connectivity of area/portal graph.
Since we force the shadows from all world geometry, it is logical to do the same for portal-closing doors as well.

git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10642 49c82d7f-2e2a-0410-a16f-ae8f201b507f