#6489. Block mission complete/failed events when mission is already over.
There is quite a lot of logic on mission end.
Some of it should not be run twice because end results is bad (e.g. loot stats doubled).
Some of it results in crash if run second time (in particular, inventory management crashes).
Now the first win or loss takes over, and you cannot change this state afterwards.

git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10646 49c82d7f-2e2a-0410-a16f-ae8f201b507f