#6480. Fixed visportal precision issues by removing rounding in .proc file.
Previously, all floats were rounded to integer if they were within 0.001 of it.
This caused normal vectors of BSP planes to become non-normalized and to deviate from visportal windings by a lot.
As the result, standing near some portals caused half the world to disappear.

Now floats are printed as integers only if the two match exactly.
This fixes problems in Dragon's Claw and Return to The City missions.
The old behavior can be restored by setting dmap_outputNoSnap to 0.


git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10654 49c82d7f-2e2a-0410-a16f-ae8f201b507f