#5600. Disabled dynamic creation of unknown/unregistered cvars.
It was previously possible to create a cvar without registering it statically as: cvarSystem->SetCVarFloat("myNewCvar", 1.5);
However, there are no uses for such cvars currently: user can see that they exist, and that's all.

The problem with such a possibility is that when new cvar is created, no other queries by string name can be done in parallel.
We have quite a lot string name queries across the code (like cvarSystem->GetCVarBool("r_swapInterval")), so it is much easier to block cvars creation during parallel sections.
In fact, the only moment when it legimitely happens now is reading .cfg files, which happens during initialization (which is strictly sequental).

If player switches platform or TDM version and uses the old .cfg file, then it often contains some unknown cvars which are not registered.
Previously they were created dynamically, but they were not saved to .cfg file afterwards.
Since they were not persisten anyway, there is no point in creating them in the first place: and now we don't create them.

Moreover, trying to create new cvar dynamically after initialization now results in a warning (better fix such code if it exists).
Finally, I moved initialization of jobs system later in the sequence, when cvars creation is already forbidden.

git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10716 49c82d7f-2e2a-0410-a16f-ae8f201b507f