#5718. Replaced clumsy self-written bubble sort with std::stable_sort in idMaterial::SortInteractionStages.
Note: this does not change behavior, but will make it easier for the future adventurers to understand what's going on.

The old algorithm does compare-and-swap for consecutive elements only.
It does full sweep from left to right, which results in the maximum element getting to last place.
Then it does a second sweep without toughing the last element, which moves the second-maximum elements to the pre-last place.
And it goes on and on.
As the result, we get sorted array.

If the array contains equal elements, then their relative order does not change.
This is easy to verify because we only swap consecutive elements, and we don't swap them when they are equal.


git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10764 49c82d7f-2e2a-0410-a16f-ae8f201b507f