#6546. Added idRenderModel::GetLoadVersion method to detect model reloads.
Normally, a model is immutable/const.
So if we computed some derived data for it, it should stay valid indefinitely.
However, doing "reloadModels" can substitute the contents of changed models with different data, and then the derived data should be recomputed.
The idea of load version is that whoever keeps derived data can detect modification and drop & recompute it.

I'll use this for light estimate system.
For now it only makes sense for static models (MD5 models can't be reloaded at all).

git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10820 49c82d7f-2e2a-0410-a16f-ae8f201b507f