Added custom range compressing curve to tonemap (optional), and also "exposure" multiplier (optional too).
This is stimulated by the topic:
  https://forums.thedarkmod.com/index.php?/topic/22600-tone-mapping
Moreover, I compared the curve to a sample integration of AgX:
  https://forums.thedarkmod.com/index.php?/topic/22600-tone-mapping/#findComment-498493

The reasoning and math behind this implementation is dscribed here:
  https://colab.research.google.com/gist/stgatilov/640485ffb49fb734e0642f6d5e34dff8/tonemap_compression_curve.ipynb
Basically, it is "almost linear" at the beginning (exponential reduction to lower derivative at switch point), and customized Reinhard at the tail, joined smoothly at switch point.

The new curve is yet disabled by default, but can be enabled with "r_postprocess_compress 1".
There are also a bunch of cvars starting with same string, which can be used to tweak the curve.

Also note that the current "brightness" is an awful thing since it breaks output range.
Instead it's better to use a multiplier BEFORE compression, which is now available as r_postprocess_exposure.


git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10912 49c82d7f-2e2a-0410-a16f-ae8f201b507f