#6610. Properly skip commented lines in OBJ, don't stop on syntax error in comments.
It turns out that idLexer::SkipRestOfLine simply reads tokens until EOL is met.
So if there is something tokenizer does not like in skipped text, it producer error.

On the other hand, there was idLexer::ReadRestOfLine, which did not tokenize the read section.
I reworked it into ReadRestOfLineUnp, which can return the string or not (also read past EOL or not).
The usages of ReadRestOfLine replaced: change in behavior is not important here, the only valid usage is mapsequence.txt for campaigns.


git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10970 49c82d7f-2e2a-0410-a16f-ae8f201b507f