Fixed stack overflow on laarge static model.
The discussion is here:
  https://forums.thedarkmod.com/index.php?/topic/22672-214-replace-malloca16-calls-with-mem_malloca/

Note that I use our idFlexList<> here instead of more low-level malloc/alloca hybrid that is used in dhwem3.
The main benefit of this is that the contents are easily viewably in the debugger (thanks to natvis).
The downside may be slightly more code bloat and lower limit for switching to heap memory, but I think we can handle it.

git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@10993 49c82d7f-2e2a-0410-a16f-ae8f201b507f